Move Custom Logger into Utils
