+++ lib/connection-manager.js
@@ -10,7 +10,6 @@ const url = require('url');
 const fileUtils = require('./file-utilities');
 const getProxyFromURI = require('./getProxyFromURI');
 const linebuffer = require('./linebuffer');
-const log = require('./log');
 const utils = require('./utils');
 const winUtils = require('./windows-utilities');
 
@@ -47,22 +46,22 @@ const checkHealth = (config) => {
             });
         }, (error, body) => {
             if (error) {
-                log(`Error from HealthCheck Endpoint: ${error}`);
+                utils.log(`Error from HealthCheck Endpoint: ${error}`);
             } else if (body && body.statusCode === 200) {
                 if (body.autoupdate === true) {
                     if (config.autoupdate !== 0) {
-                        log('LogDNA Agent Update Request Received - Auto Updating...');
+                        utils.log('LogDNA Agent Update Request Received - Auto Updating...');
                         return utils.updateSelf();
                     }
                 } else if (body.restart === true) {
-                    log('LogDNA Agent Restart Request Received - Restarting...');
+                    utils.log('LogDNA Agent Restart Request Received - Restarting...');
                     return utils.restartSelf();
                 }
                 if (body.endpoint) {
                     config.LOGDNA_LOGURL = config.LOGDNA_LOGURL.replace(config.LOGDNA_LOGHOST, body.endpoint);
                 }
             } else {
-                log(`Response from HealthCheck Endpoint: ${JSON.stringify(body)}`);
+                utils.log(`Response from HealthCheck Endpoint: ${JSON.stringify(body)}`);
             }
 
             if (firstrun) {
@@ -75,7 +74,7 @@ const checkHealth = (config) => {
                 }
                 firstrun = false;
             } else {
-                log('Streaming Resumed after HealthCheck Call');
+                utils.log('Streaming Resumed after HealthCheck Call');
             }
 
             return setTimeout(() => {
