+++ lib/file-utilities.js
@@ -7,7 +7,7 @@ const os = require('os');
 const spawn = require('child_process').spawn;
 
 // Internal Modules
-const log = require('./log');
+const log = require('./utils').log;
 const linebuffer = require('./linebuffer');
 const TailReadStream = require('./tailreadstream/tailreadstream');
 const Splitter = require('./tailreadstream/line-splitter');
