+++ lib/k8s.js
@@ -1,12 +1,17 @@
+// External Modules
 const debug = require('debug')('logdna:k8s');
 const net = require('net');
-const log = require('./log');
+
+// Internal Modules
+const log = require('./utils').log;
 const config = require('./config');
 
+// Constants
 const HTTP_HEADERS = 'HTTP/1.1\r\nHost: localhost\r\n\r\n';
 const LABEL_BLACKLIST = ['controller-revision-hash', 'image', 'integration-test', 'name', 'pod-template-generation', 'pod-template-hash'];
 const SANDBOX_NAME_KEY = 'io.kubernetes.sandbox.id';
 
+// Variables
 var client;
 var containerIdToNetworkId = {};
 var networkIdToLabels = {};
