+++ lib/linebuffer.js
@@ -5,7 +5,7 @@ const Queue = require('file-queue').Queue;
 const zlib = require('zlib');
 
 // Internal Modules
-const log = require('./log');
+const log = require('./utils').log;
 
 // Variables
 var buf = [];
