--- lib/log.js
@@ -1,3 +1,0 @@
-module.exports = (msg) => {
-    console.log('[' + new Date().toISOString().substring(2, 19).replace('T', ' ').replace(/[Z-]/g, '') + '] ' + msg);
-};
