+++ lib/os-version.js
@@ -1,11 +1,10 @@
 // External Modules
-const debug = require('debug')('logdna:lib:os-version');
 const fs = require('fs');
 const os = require('os');
 
 // Internal Modules
 const getOS = require('./getos');
-const log = require('./log');
+const log = require('./utils').log;
 
 // Constants
 const DEFAULT_OS_PATH = '/etc/os-release';
