+++ lib/utils.js
@@ -182,12 +182,18 @@ const appender = (xs) => {
     };
 };
 
-exports.unsetConfig = unsetConfig;
-exports.processOption = processOption;
-exports.stringify = stringify;
-exports.split = split;
+// Custom Logger
+const log = (msg) => {
+    console.log('[' + new Date().toISOString().substring(2, 19).replace('T', ' ').replace(/[Z-]/g, '') + '] ' + msg);
+};
+
+exports.appender = appender;
+exports.log = log;
 exports.pick2list = pick2list;
-exports.updateSelf = updateSelf;
+exports.processOption = processOption;
 exports.restartSelf = restartSelf;
 exports.saveConfig = saveConfig;
-exports.appender = appender;
+exports.split = split;
+exports.stringify = stringify;
+exports.unsetConfig = unsetConfig;
+exports.updateSelf = updateSelf;
