+++ lib/windows-utilities.js
@@ -2,7 +2,7 @@
 const debug = require('debug')('logdna:lib:windows-utilities');
 
 // Internal Modules
-const log = require('./log');
+const log = require('./utils').log;
 const linebuffer = require('./linebuffer');
 const WinEventLogProcessor = require('./welp');
